+++ app/jobs/hyku_addons/validate_importer_entry_job.rb
@@ -5,7 +5,7 @@ module HykuAddons
     # non_tenant_job
     def perform(account, entry, source_service_options = {}, destination_service_options = {})
       AccountElevator.switch! account.cname
-      s = HykuAddons::EntryValidationService.new(account, entry, source_service_options, destination_service_options)
+      s = HykuAddons::Validations::SolrEntryValidationService.new(account, entry, source_service_options, destination_service_options)
       s.validate
     end
   end
