+++ app/services/hyku_addons/validations/csv_entry_validation_service.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+module HykuAddons
+  module Validations
+    class CsvEntryValidationService < EntryValidationService
+      def destination_metadata
+        @_destination_metadata ||= @entry.clone.build_export_metadata
+      end
+    end
+  end
+end
